% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSD.R
\name{SSD}
\alias{SSD}
\title{Supervised Sparse Decomposition (SSD) Class}
\description{
Supervised Sparse Decomposition (SSD) Class

Supervised Sparse Decomposition (SSD) Class
}
\details{
This class implements Supervised Sparse Decomposition (SSD) for supervised feature extraction.
}
\section{Methods}{

\itemize{
\item \code{initialize(h, lambda, nepoch, beta)}: Initializes the SSD model with given parameters.
\item \code{fit(X, L, D)}: Fits the SSD model to the input data X and target L, optionally with dictionary matrix D.
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{X}}{Input matrix (m x N), feature dim: m, sample dim: N.}

\item{\code{L}}{Label vector (1 x N).}

\item{\code{D}}{Dictionary matrix, initialized or learned.}

\item{\code{Z}}{Coefficients matrix for sparse coding.}

\item{\code{W}}{Weight matrix for feature extraction.}

\item{\code{G}}{Diagonal matrix controlling sparsity in decomposition.}

\item{\code{A}}{Regression matrix for supervised learning.}

\item{\code{h}}{Dimensionality of the latent space.}

\item{\code{lambda}}{Regularization parameter for decomposition.}

\item{\code{nepoch}}{Number of epochs for model training.}

\item{\code{beta}}{Regularization parameter for supervised learning.}

\item{\code{recon_err_ra}}{Ratio of approximation error in the input space.}

\item{\code{Z_approx}}{Approximation of matrix Z.}

\item{\code{reg_mse}}{Mean squared error for supervised regression.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Supervised Sparse Decomposition-new}{\code{SSD$new()}}
\item \href{#method-Supervised Sparse Decomposition-fit}{\code{SSD$fit()}}
\item \href{#method-Supervised Sparse Decomposition-clone}{\code{SSD$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Supervised Sparse Decomposition-new"></a>}}
\if{latex}{\out{\hypertarget{method-Supervised Sparse Decomposition-new}{}}}
\subsection{Method \code{new()}}{
Initializes an SSD object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SSD$new(h, lambda, nepoch = 1000, beta = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{h}}{Integer, dimensionality of the latent space.}

\item{\code{lambda}}{Numeric, regularization parameter.}

\item{\code{nepoch}}{Integer, number of epochs for training.}

\item{\code{beta}}{Numeric, regularization parameter for supervised learning.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Supervised Sparse Decomposition-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Supervised Sparse Decomposition-fit}{}}}
\subsection{Method \code{fit()}}{
Fits the SSD model to input data and target matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SSD$fit(X, L, D = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Matrix, input data for decomposition.}

\item{\code{L}}{Matrix, label vector for supervised decomposition.}

\item{\code{D}}{Matrix, optional dictionary matrix. If NULL, D will be learned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The fitted SSD model object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Supervised Sparse Decomposition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Supervised Sparse Decomposition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SSD$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
